+++ src/pattan_sendgrid/exceptions.py
